backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-search-results==2.4.2
python-dotenv==1.0.0
arxiv
